fixed compiler warning (comparison between signed and unsigned integer expressions) as introduced in revision 22720
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@22723 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_immersed_problem/immersed_partitioned_fsi_dirichletneumann_membrane.cpp
@@ -204,7 +204,7 @@ Teuchos::RCP<Epetra_Vector> IMMERSED::ImmersedPartitionedFSIDirichletNeumannMemb
     params.set<int>("action",FLD::least_squares_matrix_rhs_immersed_boundary);
 
     // loop over all immersed fluid elements
-    for (int i=0; i<immersedbdrygids.size(); ++i)
+    for (int i=0; i<(int)immersedbdrygids.size(); ++i)
     {
       int eid = immersedbdrygids[i];
